Reorganize monorepo scripts
