+++ packages/scripts/flow-translator/rewrite-imports.js
@@ -5,7 +5,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
- * @flow
+ *
  */
 
 const fs = require('fs').promises;
