+++ packages/shared/package.json
@@ -24,8 +24,8 @@
     "@babel/preset-flow": "^7.25.9",
     "@babel/preset-react": "^7.26.3",
     "@babel/preset-typescript": "^7.26.0",
-    "@stylexjs/scripts": "0.11.1",
-    "babel-plugin-syntax-hermes-parser": "^0.26.0"
+    "babel-plugin-syntax-hermes-parser": "^0.26.0",
+    "scripts": "0.11.1"
   },
   "jest": {
     "snapshotFormat": {
